+++ packages/core/src/text/index.ts
@@ -1,6 +1,7 @@
 export * from './editor'
 export * from './style-runs'
 export * from './direction'
+export * from './case'
 export * from './coverage'
 export * from './fonts'
 export * from './fallbacks'